Crate integration_tests_common

Re-exports

pub use constants::accounts::ALICE;

pub use constants::accounts::BOB;

pub use constants::asset_hub_kusama;

pub use constants::asset_hub_polkadot;

pub use constants::asset_hub_westend;

pub use constants::bridge_hub_kusama;

pub use constants::bridge_hub_polkadot;

pub use constants::bridge_hub_rococo;

pub use constants::collectives;

pub use constants::kusama;

pub use constants::penpal;

pub use constants::polkadot;

pub use constants::rococo;

pub use constants::westend;

pub use constants::PROOF_SIZE_THRESHOLD;

pub use constants::REF_TIME_THRESHOLD;

pub use impls::RococoWococoMessageHandler;

pub use impls::WococoRococoMessageHandler;

pub use lazy_static;

pub use paste;

pub use xcm::prelude::AccountId32;

pub use xcm::prelude::All;

pub use xcm::prelude::BuyExecution;

pub use xcm::prelude::DepositAsset;

pub use xcm::prelude::RefundSurplus;

pub use xcm::prelude::Transact;

pub use xcm::prelude::UnpaidExecution;

pub use xcm::prelude::WithdrawAsset;

pub use xcm::prelude::X1;

Modules

constants

impls

sr25519

Simple sr25519 (Schnorr-Ristretto) API.

Macros

impl_accounts_helpers_for_parachain

impl_accounts_helpers_for_relay_chain

impl_assert_events_helpers_for_parachain

impl_assert_events_helpers_for_relay_chain

impl_assets_helpers_for_parachain

impl_hrmp_channels_helpers_for_relay_chain

Structs

AssetHubKusama

AssetHubKusamaReceiver

AssetHubKusamaSender

AssetHubPolkadot

AssetHubPolkadotReceiver

AssetHubPolkadotSender

AssetHubRococo

AssetHubRococoReceiver

AssetHubRococoSender

AssetHubWestend

AssetHubWestendReceiver

AssetHubWestendSender

AssetHubWococo

AssetHubWococoReceiver

AssetHubWococoSender

BridgeHubKusama

BridgeHubKusamaReceiver

BridgeHubKusamaSender

BridgeHubPolkadot

BridgeHubPolkadotReceiver

BridgeHubPolkadotSender

BridgeHubRococo

BridgeHubRococoReceiver

BridgeHubRococoSender

BridgeHubWococo

BridgeHubWococoReceiver

BridgeHubWococoSender

Collectives

CollectivesReceiver

CollectivesSender

DoubleEncoded

Wrapper around the encoded and decoded versions of a value. Caches the decoded value once computed.

GLOBAL_EXT_ASSETHUBKUSAMA

GLOBAL_EXT_ASSETHUBPOLKADOT

GLOBAL_EXT_ASSETHUBROCOCO

GLOBAL_EXT_ASSETHUBWESTEND

GLOBAL_EXT_ASSETHUBWOCOCO

GLOBAL_EXT_BRIDGEHUBKUSAMA

GLOBAL_EXT_BRIDGEHUBPOLKADOT

GLOBAL_EXT_BRIDGEHUBROCOCO

GLOBAL_EXT_BRIDGEHUBWOCOCO

GLOBAL_EXT_COLLECTIVES

GLOBAL_EXT_KUSAMA

GLOBAL_EXT_PENPALKUSAMAA

GLOBAL_EXT_PENPALKUSAMAB

GLOBAL_EXT_PENPALPOLKADOTA

GLOBAL_EXT_PENPALPOLKADOTB

GLOBAL_EXT_PENPALROCOCOA

GLOBAL_EXT_PENPALWESTENDA

GLOBAL_EXT_POLKADOT

GLOBAL_EXT_ROCOCO

GLOBAL_EXT_WESTEND

GLOBAL_EXT_WOCOCO

Kusama

KusamaMockNet

KusamaReceiver

KusamaSender

MultiAsset

Either an amount of a single fungible asset, or a single well-identified non-fungible asset.

MultiAssets

A Vec of MultiAssets.

MultiLocation

A relative path between state-bearing consensus systems.

PenpalKusamaA

PenpalKusamaAReceiver

PenpalKusamaASender

PenpalKusamaB

PenpalKusamaBReceiver

PenpalKusamaBSender

PenpalPolkadotA

PenpalPolkadotAReceiver

PenpalPolkadotASender

PenpalPolkadotB

PenpalPolkadotBReceiver

PenpalPolkadotBSender

PenpalRococoA

PenpalRococoAReceiver

PenpalRococoASender

PenpalWestendA

PenpalWestendAReceiver

PenpalWestendASender

Polkadot

PolkadotMockNet

PolkadotReceiver

PolkadotSender

Rococo

RococoMockNet

RococoReceiver

RococoSender

RococoWococoMockBridge

Storage

Struct containing data needed for a storage.

Weight

Westend

WestendMockNet

WestendReceiver

WestendSender

Wococo

WococoMockNet

WococoReceiver

WococoRococoMockBridge

WococoSender

Xcm

Enums

AggregateMessageOrigin

Aggregate message origin for the MessageQueue pallet.

Error

Error codes used in XCM. The first errors codes have explicit indices and are part of the XCM format. Those trailing are merely part of the XCM implementation; there is no expectation that they will retain the same index over time.

OriginKind

Basically just the XCM (more general) version of ParachainDispatchOrigin.

Outcome

Outcome of an XCM execution.

UmpQueueId

Identifies a UMP queue inside the MessageQueue pallet.

VersionedXcm

A single XCM message, together with its version code.

WeightLimit

An optional weight limit.

Constants

LOCAL_EXT_ASSETHUBKUSAMA

LOCAL_EXT_ASSETHUBPOLKADOT

LOCAL_EXT_ASSETHUBROCOCO

LOCAL_EXT_ASSETHUBWESTEND

LOCAL_EXT_ASSETHUBWOCOCO

LOCAL_EXT_BRIDGEHUBKUSAMA

LOCAL_EXT_BRIDGEHUBPOLKADOT

LOCAL_EXT_BRIDGEHUBROCOCO

LOCAL_EXT_BRIDGEHUBWOCOCO

LOCAL_EXT_COLLECTIVES

LOCAL_EXT_KUSAMA

LOCAL_EXT_PENPALKUSAMAA

LOCAL_EXT_PENPALKUSAMAB

LOCAL_EXT_PENPALPOLKADOTA

LOCAL_EXT_PENPALPOLKADOTB

LOCAL_EXT_PENPALROCOCOA

LOCAL_EXT_PENPALWESTENDA

LOCAL_EXT_POLKADOT

LOCAL_EXT_ROCOCO

LOCAL_EXT_WESTEND

LOCAL_EXT_WOCOCO

Traits

AssetHubKusamaPallet

AssetHubPolkadotPallet

AssetHubRococoPallet

AssetHubWestendPallet

AssetHubWococoPallet

BridgeHubKusamaPallet

BridgeHubPolkadotPallet

BridgeHubRococoPallet

BridgeHubWococoPallet

CollectivesPallet

Encode

Trait that allows zero-copy write of value-references to slices in LE format.

Get

A trait for querying a single value from a type.

KusamaPallet

PenpalKusamaAPallet

PenpalKusamaBPallet

PenpalPolkadotAPallet

PenpalPolkadotBPallet

PenpalRococoAPallet

PenpalWestendAPallet

PolkadotPallet

RococoPallet

WestendPallet

WococoPallet

Functions

xcm_transact_paid_execution

Helper method to build a XCM with a Transact instruction and paying for its execution

xcm_transact_unpaid_execution

Helper method to build a XCM with a Transact instruction without paying for its execution

Type Aliases

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Balance

Balance of an account.

Derive Macros

Encode

Derive parity_scale_codec::Encode and parity_scale_codec::EncodeLike for struct and enum.