Crate integration_tests_common

Re-exports

pub use constants::accounts::ALICE;

pub use constants::accounts::BOB;

pub use constants::bridge_hub_kusama;

pub use constants::bridge_hub_polkadot;

pub use constants::collectives;

pub use constants::kusama;

pub use constants::penpal;

pub use constants::polkadot;

pub use constants::statemine;

pub use constants::statemint;

Modules

constants

sr25519

Simple sr25519 (Schnorr-Ristretto) API.

Structs

BHKusama

BHKusamaReceiver

BHKusamaSender

BHPolkadot

BHPolkadotReceiver

BHPolkadotSender

Collectives

CollectivesReceiver

CollectivesSender

Kusama

KusamaMockNet

KusamaReceiver

KusamaSender

PenpalKusama

PenpalKusamaReceiver

PenpalKusamaSender

PenpalPolkadot

PenpalPolkadotReceiver

PenpalPolkadotSender

Polkadot

PolkadotMockNet

PolkadotReceiver

PolkadotSender

Statemine

StatemineReceiver

StatemineSender

Statemint

StatemintReceiver

StatemintSender

Storage

Struct containing data needed for a storage.

Constants

EXT_BHKUSAMA

EXT_BHPOLKADOT

EXT_COLLECTIVES

EXT_KUSAMA

EXT_PENPALKUSAMA

EXT_PENPALPOLKADOT

EXT_POLKADOT

EXT_STATEMINE

EXT_STATEMINT

Traits

BHKusamaPallet

BHPolkadotPallet

CollectivesPallet

Get

A trait for querying a single value from a type.

KusamaPallet

PenpalKusamaPallet

PenpalPolkadotPallet

PolkadotPallet

StateminePallet

StatemintPallet

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AuraId

Balance

Balance of an account.

BlockNumber

An index to a block.

StatemintAuraId