Crate integration_tests_common

Re-exports

pub use constants::accounts::ALICE;

pub use constants::accounts::BOB;

pub use constants::asset_hub_kusama;

pub use constants::asset_hub_polkadot;

pub use constants::asset_hub_westend;

pub use constants::bridge_hub_kusama;

pub use constants::bridge_hub_polkadot;

pub use constants::bridge_hub_rococo;

pub use constants::collectives;

pub use constants::kusama;

pub use constants::penpal;

pub use constants::polkadot;

pub use constants::rococo;

pub use constants::westend;

pub use impls::RococoWococoMessageHandler;

pub use impls::WococoRococoMessageHandler;

Modules

constants

impls

sr25519

Simple sr25519 (Schnorr-Ristretto) API.

Structs

AssetHubKusama

AssetHubKusamaReceiver

AssetHubKusamaSender

AssetHubPolkadot

AssetHubPolkadotReceiver

AssetHubPolkadotSender

AssetHubRococo

AssetHubRococoReceiver

AssetHubRococoSender

AssetHubWestend

AssetHubWestendReceiver

AssetHubWestendSender

AssetHubWococo

AssetHubWococoReceiver

AssetHubWococoSender

BridgeHubKusama

BridgeHubKusamaReceiver

BridgeHubKusamaSender

BridgeHubPolkadot

BridgeHubPolkadotReceiver

BridgeHubPolkadotSender

BridgeHubRococo

BridgeHubRococoReceiver

BridgeHubRococoSender

BridgeHubWococo

BridgeHubWococoReceiver

BridgeHubWococoSender

Collectives

CollectivesReceiver

CollectivesSender

Kusama

KusamaMockNet

KusamaReceiver

KusamaSender

PenpalKusama

PenpalKusamaReceiver

PenpalKusamaSender

PenpalPolkadot

PenpalPolkadotReceiver

PenpalPolkadotSender

PenpalWestend

PenpalWestendReceiver

PenpalWestendSender

Polkadot

PolkadotMockNet

PolkadotReceiver

PolkadotSender

Rococo

RococoMockNet

RococoReceiver

RococoSender

RococoWococoMockBridge

Storage

Struct containing data needed for a storage.

Westend

WestendMockNet

WestendReceiver

WestendSender

Wococo

WococoMockNet

WococoReceiver

WococoRococoMockBridge

WococoSender

Constants

EXT_ASSETHUBKUSAMA

EXT_ASSETHUBPOLKADOT

EXT_ASSETHUBROCOCO

EXT_ASSETHUBWESTEND

EXT_ASSETHUBWOCOCO

EXT_BRIDGEHUBKUSAMA

EXT_BRIDGEHUBPOLKADOT

EXT_BRIDGEHUBROCOCO

EXT_BRIDGEHUBWOCOCO

EXT_COLLECTIVES

EXT_KUSAMA

EXT_PENPALKUSAMA

EXT_PENPALPOLKADOT

EXT_PENPALWESTEND

EXT_POLKADOT

EXT_ROCOCO

EXT_WESTEND

EXT_WOCOCO

Traits

AssetHubKusamaPallet

AssetHubPolkadotPallet

AssetHubRococoPallet

AssetHubWestendPallet

AssetHubWococoPallet

BridgeHubKusamaPallet

BridgeHubPolkadotPallet

BridgeHubRococoPallet

BridgeHubWococoPallet

CollectivesPallet

Get

A trait for querying a single value from a type.

KusamaPallet

PenpalKusamaPallet

PenpalPolkadotPallet

PenpalWestendPallet

PolkadotPallet

RococoPallet

WestendPallet

WococoPallet

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AssetHubPolkadotAuraId

AuraId

Balance

Balance of an account.

BlockNumber

An index to a block.