Crate integration_tests_common

Re-exports

pub use constants::accounts::ALICE;

pub use constants::accounts::BOB;

pub use constants::asset_hub_kusama;

pub use constants::asset_hub_polkadot;

pub use constants::asset_hub_westend;

pub use constants::bridge_hub_kusama;

pub use constants::bridge_hub_polkadot;

pub use constants::collectives;

pub use constants::kusama;

pub use constants::penpal;

pub use constants::polkadot;

pub use constants::westend;

Modules

constants

sr25519

Simple sr25519 (Schnorr-Ristretto) API.

Structs

AssetHubKusama

AssetHubKusamaReceiver

AssetHubKusamaSender

AssetHubPolkadot

AssetHubPolkadotReceiver

AssetHubPolkadotSender

AssetHubWestend

AssetHubWestendReceiver

AssetHubWestendSender

BHKusama

BHKusamaReceiver

BHKusamaSender

BHPolkadot

BHPolkadotReceiver

BHPolkadotSender

Collectives

CollectivesReceiver

CollectivesSender

Kusama

KusamaMockNet

KusamaReceiver

KusamaSender

PenpalKusama

PenpalKusamaReceiver

PenpalKusamaSender

PenpalPolkadot

PenpalPolkadotReceiver

PenpalPolkadotSender

PenpalWestend

PenpalWestendReceiver

PenpalWestendSender

Polkadot

PolkadotMockNet

PolkadotReceiver

PolkadotSender

Storage

Struct containing data needed for a storage.

Westend

WestendMockNet

WestendReceiver

WestendSender

Constants

EXT_ASSETHUBKUSAMA

EXT_ASSETHUBPOLKADOT

EXT_ASSETHUBWESTEND

EXT_BHKUSAMA

EXT_BHPOLKADOT

EXT_COLLECTIVES

EXT_KUSAMA

EXT_PENPALKUSAMA

EXT_PENPALPOLKADOT

EXT_PENPALWESTEND

EXT_POLKADOT

EXT_WESTEND

Traits

AssetHubKusamaPallet

AssetHubPolkadotPallet

AssetHubWestendPallet

BHKusamaPallet

BHPolkadotPallet

CollectivesPallet

Get

A trait for querying a single value from a type.

KusamaPallet

PenpalKusamaPallet

PenpalPolkadotPallet

PenpalWestendPallet

PolkadotPallet

WestendPallet

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AssetHubPolkadotAuraId

AuraId

Balance

Balance of an account.

BlockNumber

An index to a block.